Move DevtoolsRPCMessage to rpc file
