+++ src/testUtils/testMessageAdapter.ts
@@ -1,10 +1,10 @@
 import type { MessageAdapter } from "../extension/messageAdapters";
 import { isRPCRequestMessage, MessageType } from "../extension/messages";
 import type {
-  DevtoolsRPCMessage,
   RPCRequestMessage,
   RPCResponseMessage,
 } from "../extension/messages";
+import type { DevtoolsRPCMessage } from "../extension/rpc";
 import type { SafeAny } from "../types";
 import { createId } from "../utils/createId";
